pyproject.toml
[tool.poetry]
name = "agentic-rag"
version = "0.1.0"
description = "Agentic RAG system using Python and Azure OpenAI"
authors = ["Orhan yazar.or@gmail.com"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.2.0"
langchain = "^0.1.0"
azure-search-documents = "^11.4.0"
faiss-cpu = "^1.7.4"
azure-identity = "^1.14.0"
python-dotenv = "^1.0.1"
tiktoken = "^0.8.0"
sentence-transformers = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "agentic_rag.main:run"